Change uint/int arg and return types

+++ roaring.go
@@ -1878,9 +1878,9 @@ func (rb *Bitmap) CloneCopyOnWriteContainers() {
 	rb.highlowcontainer.cloneCopyOnWriteContainers()
 }
 
-func (rb *Bitmap) NextValue(target int) int {
-	originalKey := highbits(uint32(target))
-	query := lowbits(uint32(target))
+func (rb *Bitmap) NextValue(target uint32) int64 {
+	originalKey := highbits(target)
+	query := lowbits(target)
 	nextValue := -1
 	containerIndex := rb.highlowcontainer.advanceUntil(originalKey, -1)
 	for containerIndex < rb.highlowcontainer.size() && nextValue == -1 {
@@ -1907,10 +1907,10 @@ func (rb *Bitmap) NextValue(target int) int {
 		containerIndex++
 	}
 
-	return nextValue
+	return int64(nextValue)
 }
 
-func (rb *Bitmap) PreviousValue(target int) int {
+func (rb *Bitmap) PreviousValue(target uint32) int64 {
 	if rb.IsEmpty() {
 		return -1
 	}
@@ -1921,7 +1921,7 @@ func (rb *Bitmap) PreviousValue(target int) int {
 	containerIndex := rb.highlowcontainer.advanceUntil(originalKey, -1)
 
 	if containerIndex == rb.highlowcontainer.size() {
-		return int(rb.Maximum())
+		return int64(rb.Maximum())
 	}
 
 	if rb.highlowcontainer.getKeyAtIndex(containerIndex) > originalKey {
@@ -1953,34 +1953,34 @@ func (rb *Bitmap) PreviousValue(target int) int {
 		containerIndex--
 	}
 
-	return prevValue
+	return int64(prevValue)
 }
 
-func (rb *Bitmap) NextAbsentValue(target int) int {
-	originalKey := highbits(uint32(target))
-	query := lowbits(uint32(target))
+func (rb *Bitmap) NextAbsentValue(target uint32) int64 {
+	originalKey := highbits(target)
+	query := lowbits(target)
 	nextValue := -1
 
 	containerIndex := rb.highlowcontainer.advanceUntil(originalKey, -1)
 	if containerIndex == rb.highlowcontainer.size() {
 		// if we are here it means no container found, just return the target
-		return target
+		return int64(target)
 	}
 
 	containerKey := rb.highlowcontainer.getKeyAtIndex(containerIndex)
 
 	keyspace := uint32(containerKey) << 16
-	if target < int(keyspace) {
+	if target < keyspace {
 		// target is less than the start of the keyspace start
 		// that means target cannot be in the keyspace
-		return target
+		return int64(target)
 	}
 
 	container := rb.highlowcontainer.getContainer(containerKey)
 	nextValue = container.nextAbsentValue(query)
 	for {
 		if nextValue != (1 << 16) {
-			return int(combineLoHi32(uint32(nextValue), keyspace))
+			return int64(combineLoHi32(uint32(nextValue), keyspace))
 		}
 
 		if containerIndex == rb.highlowcontainer.size()-1 {
@@ -1988,14 +1988,14 @@ func (rb *Bitmap) NextAbsentValue(target int) int {
 			if err == nil {
 				return -1
 			}
-			return int(val) + 1
+			return int64(val) + 1
 		}
 		containerIndex++
 		nextContainerKey := rb.highlowcontainer.getKeyAtIndex(containerIndex)
 		if containerKey < nextContainerKey {
 			// There is a gap between keys
 			// Just increment the current key and shift to get HoB
-			return int(containerKey+1) << 16
+			return int64(containerKey+1) << 16
 		}
 		containerKey = nextContainerKey
 		container = rb.highlowcontainer.getContainer(containerKey)
@@ -2003,36 +2003,36 @@ func (rb *Bitmap) NextAbsentValue(target int) int {
 	}
 }
 
-func (rb *Bitmap) PreviousAbsentValue(target int) int {
-	originalKey := highbits(uint32(target))
-	query := lowbits(uint32(target))
+func (rb *Bitmap) PreviousAbsentValue(target uint32) int64 {
+	originalKey := highbits(target)
+	query := lowbits(target)
 	prevValue := -1
 
 	containerIndex := rb.highlowcontainer.advanceUntil(originalKey, -1)
 
 	if containerIndex == rb.highlowcontainer.size() {
 		// if we are here it means no container found, just return the target
-		return target
+		return int64(target)
 	}
 
 	if containerIndex == -1 {
 		// if we are here it means no container found, just return the target
-		return target
+		return int64(target)
 	}
 
 	containerKey := rb.highlowcontainer.getKeyAtIndex(containerIndex)
 	keyspace := uint32(containerKey) << 16
-	if target < int(keyspace) {
+	if target < keyspace {
 		// target is less than the start of the keyspace start
 		// that means target cannot be in the keyspace
-		return target
+		return int64(target)
 	}
 
 	container := rb.highlowcontainer.getContainer(containerKey)
 	prevValue = container.previousAbsentValue(query)
 	for {
 		if prevValue != -1 {
-			return int(combineLoHi32(uint32(prevValue), keyspace))
+			return int64(combineLoHi32(uint32(prevValue), keyspace))
 		}
 
 		if containerIndex == 0 {
@@ -2041,14 +2041,14 @@ func (rb *Bitmap) PreviousAbsentValue(target int) int {
 				// OR panic, Java panics
 				return -1
 			}
-			return int(val) - 1
+			return int64(val) - 1
 		}
 		containerIndex--
 		nextContainerKey := rb.highlowcontainer.getKeyAtIndex(containerIndex)
 		if nextContainerKey < containerKey-1 {
 			// There is a gap between keys, eg missing container
 			// Just decrement the current key and shift to get HoB of the missing container
-			return (int(containerKey) << 16) - 1
+			return (int64(containerKey) << 16) - 1
 		}
 		containerKey = nextContainerKey
 		container = rb.highlowcontainer.getContainer(containerKey)
